+++ src/DataFixtures/AppFixtures.php
@@ -6,6 +6,7 @@ use App\Entity\ApiUser;
 use App\Entity\Screen;
 use App\Entity\Tags;
 use App\Entity\Tile;
+use App\Repository\ApiUserRepository;
 use App\Repository\TagsRepository;
 use Doctrine\Bundle\FixturesBundle\Fixture;
 use Doctrine\Persistence\ObjectManager;
@@ -13,7 +14,8 @@ use Doctrine\Persistence\ObjectManager;
 class AppFixtures extends Fixture
 {
     public function __construct(
-       private readonly TagsRepository $tagsRepository
+       private readonly TagsRepository $tagsRepository,
+       private readonly ApiUserRepository $apiUserRepository,
     ) {
     }
 
@@ -24,6 +26,17 @@ class AppFixtures extends Fixture
      */
     public function load(ObjectManager $manager): void
     {
+        // API user
+        $json = \file_get_contents(__DIR__.'/data/api_user.json');
+        $data = \json_decode($json, false, 512, JSON_THROW_ON_ERROR);
+        foreach ($data as $datum) {
+            $api = new ApiUser();
+            $api->setName($datum->name)
+                ->setToken($datum->token)
+                ->setRemoteApiKey($datum->remoteApiKey);
+            $manager->persist($api);
+        }
+
         // Tiles
         $json = \file_get_contents(__DIR__.'/data/tiles.json');
         $data = \json_decode($json, false, 512, JSON_THROW_ON_ERROR);
@@ -49,6 +62,7 @@ class AppFixtures extends Fixture
         }
 
         // Screens
+        $apiUser = $this->apiUserRepository->findOneBy(['token' => '123456789']);
         $json = \file_get_contents(__DIR__.'/data/screen.json');
         $data = \json_decode($json, false, 512, JSON_THROW_ON_ERROR);
         foreach ($data as $datum) {
@@ -56,7 +70,8 @@ class AppFixtures extends Fixture
             $screen->setTitle($datum->title)
                 ->setGridColumns($datum->gridColumns)
                 ->setGridRows($datum->gridRows)
-                ->setVariant(json_encode($datum->variant));
+                ->setVariant(json_encode($datum->variant))
+                ->setApiUser($apiUser);
 
             foreach ($datum->tags as $tag) {
                 $entity = $this->createTags($manager, $tag);
@@ -64,16 +79,6 @@ class AppFixtures extends Fixture
             }
 
             $manager->persist($screen);
-        }
-
-        // API user
-        $json = \file_get_contents(__DIR__.'/data/api_user.json');
-        $data = \json_decode($json, false, 512, JSON_THROW_ON_ERROR);
-        foreach ($data as $datum) {
-            $api = new ApiUser();
-            $api->setToken($datum->token)
-                ->setRemoteApiKey($datum->remoteApiKey);
-            $manager->persist($api);
         }
 
         $manager->flush();