+++ src/Entity/ApiUser.php
@@ -3,6 +3,8 @@
 namespace App\Entity;
 
 use App\Repository\ApiUserRepository;
+use Doctrine\Common\Collections\ArrayCollection;
+use Doctrine\Common\Collections\Collection;
 use Doctrine\ORM\Mapping as ORM;
 use Gedmo\Timestampable\Traits\TimestampableEntity;
 use Symfony\Component\Security\Core\User\UserInterface;
@@ -25,6 +27,22 @@ class ApiUser implements UserInterface
 
     #[ORM\Column(length: 255)]
     private string $remoteApiKey;
+
+    #[ORM\OneToMany(mappedBy: 'apiUser', targetEntity: Screen::class)]
+    private Collection $screens;
+
+    #[ORM\Column(length: 255)]
+    private ?string $name = null;
+
+    public function __construct()
+    {
+        $this->screens = new ArrayCollection();
+    }
+
+    public function __toString(): string
+    {
+        return $this->name ?? '';
+    }
 
     public function getId(): ?int
     {
@@ -87,6 +105,48 @@ class ApiUser implements UserInterface
     public function setRemoteApiKey(string $remoteApiKey): self
     {
         $this->remoteApiKey = $remoteApiKey;
+
+        return $this;
+    }
+
+    /**
+     * @return Collection<int, Screen>
+     */
+    public function getScreens(): Collection
+    {
+        return $this->screens;
+    }
+
+    public function addScreen(Screen $screen): self
+    {
+        if (!$this->screens->contains($screen)) {
+            $this->screens->add($screen);
+            $screen->setApiUser($this);
+        }
+
+        return $this;
+    }
+
+    public function removeScreen(Screen $screen): self
+    {
+        if ($this->screens->removeElement($screen)) {
+            // set the owning side to null (unless already changed)
+            if ($screen->getApiUser() === $this) {
+                $screen->setApiUser(null);
+            }
+        }
+
+        return $this;
+    }
+
+    public function getName(): ?string
+    {
+        return $this->name;
+    }
+
+    public function setName(string $name): self
+    {
+        $this->name = $name;
 
         return $this;
     }